+++ CHANGELOG.md
@@ -19,4 +19,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Added:
   - All the C code for the minimum viable product solver
-  - More documentation
+- Updated:
+  - README
+
+## 1.1.0 [2021-01-26]
+
+- Added:
+  - Cheating
+  - Configuration
+  - Timer-based search for a more optimal solution
+- Updated:
+  - README
+  - Solver help messages are now more explicit, and match those in the README